# packages/origine2/src/config/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/manageTemplate/templateList": {
      "get": {
        "operationId": "ManageTemplateController_getTemplateList",
        "summary": "Retrieve template list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returned template list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateInfoDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/createTemplate": {
      "post": {
        "operationId": "ManageTemplateController_createTemplate",
        "summary": "Create a new template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Template creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template creation result."
          },
          "400": {
            "description": "Invalid input data."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/getTemplateConfig/{templateDir}": {
      "get": {
        "operationId": "ManageTemplateController_getTemplateConfig",
        "summary": "Get Template Configuration",
        "parameters": [
          {
            "name": "templateDir",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned template configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateConfigDto"
                }
              }
            }
          },
          "404": {
            "description": "Template configuration not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/updateTemplateConfig": {
      "put": {
        "operationId": "ManageTemplateController_updateTemplateConfig",
        "summary": "Update template configuration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template configuration updated successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "404": {
            "description": "Template configuration not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/delete/{templateDir}": {
      "delete": {
        "operationId": "ManageTemplateController_deleteTemplate",
        "summary": "Delete Template",
        "parameters": [
          {
            "name": "templateDir",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Template deleted successfully."
          },
          "404": {
            "description": "Template not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/trash/{templateDir}": {
      "delete": {
        "operationId": "ManageTemplateController_trashTemplate",
        "summary": "Delete Template",
        "parameters": [
          {
            "name": "templateDir",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Template trashed successfully."
          },
          "404": {
            "description": "Template not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/applyTemplateToGame": {
      "post": {
        "operationId": "ManageTemplateController_applyTemplateToGame",
        "summary": "Apply template to a game",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplateToGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template applied successfully."
          },
          "400": {
            "description": "Invalid input data or failed to apply template."
          },
          "404": {
            "description": "Template or game directory not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/getStyleByClassName": {
      "post": {
        "operationId": "ManageTemplateController_getStyleByClassName",
        "summary": "Get style by class name",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetStyleByClassNameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returned style information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          },
          "404": {
            "description": "Style not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/outputTemplate": {
      "post": {
        "operationId": "ManageTemplateController_outputTemplate",
        "summary": "Output Template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutputTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Output Template Successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Output Template Failed."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    },
    "/api/manageTemplate/importTemplate": {
      "post": {
        "operationId": "ManageTemplateController_importTemplate",
        "summary": "Import Template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImportTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import Template Successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Import Template Failed."
          }
        },
        "tags": [
          "Manage Template"
        ]
      }
    }
  },
  "info": {
    "title": "WebGAL Terre API",
    "description": "API Refrence of WebGAL Terre Editor",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TemplateFontConfigDto": {
        "type": "object",
        "properties": {
          "font-family": {
            "type": "string",
            "description": "The font-family name"
          },
          "url": {
            "type": "string",
            "description": "The url of the font file"
          },
          "type": {
            "type": "string",
            "description": "The font type used for @font-face format",
            "enum": [
              "truetype",
              "opentype",
              "woff",
              "woff2",
              "embedded-opentype",
              "svg"
            ]
          }
        },
        "required": [
          "font-family",
          "url",
          "type"
        ]
      },
      "TemplateConfigDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the template"
          },
          "id": {
            "type": "string",
            "description": "The id of the template"
          },
          "webgal-version": {
            "type": "string",
            "description": "The webgal version of the template"
          },
          "fonts": {
            "description": "The font registrations of the template",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateFontConfigDto"
            }
          }
        },
        "required": [
          "name",
          "webgal-version"
        ]
      },
...
      "TemplateInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the template"
          },
          "id": {
            "type": "string",
            "description": "The id of the template"
          },
          "webgal-version": {
            "type": "string",
            "description": "The webgal version of the template"
          },
          "fonts": {
            "description": "The font registrations of the template",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateFontConfigDto"
            }
          },
          "dir": {
            "type": "string",
            "description": "The dir of the template"
          },
          "builtIn": {
            "type": "boolean",
            "description": "Whether the template is built in"
          }
        },
        "required": [
          "name",
          "webgal-version",
          "dir",
          "builtIn"
        ]
      },
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string",
            "description": "The name of the template to be created"
          },
          "templateDir": {
            "type": "string",
            "description": "The dir of the template"
          }
        },
        "required": [
          "templateName",
          "templateDir"
        ]
      },
      "UpdateTemplateConfigDto": {
        "type": "object",
        "properties": {
          "templateDir": {
            "type": "string",
            "description": "The dir of the template"
          },
          "newTemplateConfig": {
            "description": "The new config of the template",
            "allOf": [
              {
                "$ref": "#/components/schemas/TemplateConfigDto"
              }
            ]
          }
        },
        "required": [
          "templateDir",
          "newTemplateConfig"
        ]
      },
      "ApplyTemplateToGameDto": {
        "type": "object",
        "properties": {
          "templateDir": {
            "type": "string",
            "description": "The template name to apply"
          },
          "gameDir": {
            "type": "string",
            "description": "The game name to be applied."
          }
        },
        "required": [
          "templateDir",
          "gameDir"
        ]
      },
      "GetStyleByClassNameDto": {
        "type": "object",
        "properties": {
          "className": {
            "type": "string",
            "description": "The name of class to be fetched"
          },
          "filePath": {
            "type": "string",
            "description": "The path of stylesheet file to be fetched"
          }
        },
        "required": [
          "className",
          "filePath"
        ]
      },
      "OutputTemplateDto": {
        "type": "object",
        "properties": {
          "templateDir": {
            "type": "string",
            "description": "The template directory name"
          }
        },
        "required": [
          "templateDir"
        ]
      },
      "ImportTemplateDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The template's zip file"
          }
        },
        "required": [
          "file"
        ]
      }
    }
  }
}